#  ActiveModel::Attribute: elide dup for immutable types