--- a/activemodel/lib/active_model/type/string.rb
+++ b/activemodel/lib/active_model/type/string.rb
@@ -19,6 +19,10 @@ module ActiveModel
         end
       end
 
+      def mutable? # :nodoc:
+        true
+      end
+
       def to_immutable_string
         ImmutableString.new(
           true: @true,