@george-ma 2025-01-21
had a concern about this PR Jean, if you have a sec to give your thoughts cc: @byroot https://github.com/Shopify/shopify/pull/569940#issuecomment-2605036320